🧹 Separated Job accessors from Job Status accessors
